Added --delete-dst-snapshots-no-crosscheck CLI option.
This flag indicates that --delete-dst-snapshots=snapshots shall check the source dataset only for snapshots with the same GUID, and ignore whatever bookmarks may be present in the source dataset. Similarly, it also indicates that --delete-dst-snapshots=bookmarks shall check the source dataset only for bookmarks with the same GUID, and ignore whatever snapshots may be present in the source dataset.

Signed-off-by: Wolfgang Hoschek <wolfgang.hoschek@mac.com>