--include-snapshot-times-and-ranks: Add 'anytime' and 'notime' keywords.
Add 'anytime' as a more user-friendly filter syntax to indicate unbounded start time and unbounded end time. 'anytime' is an alias for the existing '*' wildcard character.

Also add 'notime' keyword as a more user-friendly filter syntax to indicate an empty time range. 'notime' is an alias for the existing '0..0'.

Examples:
'anytime..7 days ago' is the same as '*..7 days ago'
'7 days ago..anytime' is the same as '7 days ago..*'
'notime' is the same as '0..0'

Signed-off-by: Wolfgang Hoschek <wolfgang.hoschek@mac.com>