Add --create-src-snapshot-* CLI options for efficiently creating periodic (and adhoc) atomic snapshots of datasets
The name of the snapshot can be configured via --create-src-snapshot-* suboptions (see below). To create snapshots only, without any other processing such as replication, etc, consider using this flag together with the --skip-replication flag.

Typically, a `cron` job on the source host runs `bzfs` periodically to create new snapshots and prune outdated snapshots on the source, whereas another `cron` job on the destination host runs `bzfs` periodically to prune outdated destination snapshots. Yet another `cron` job runs `bzfs` periodically to replicate the recently created snapshots from the source to the destination. The frequency of these periodic activities is typically every second, minute, hour, day, week, month and/or year (or multiples thereof).

A periodic snapshot is created if it is due per the schedule indicated by its --create-src-snapshot-suffix (e.g. '_daily' or '_hourly' or _'10minutely' or '_2secondly'), or if the --create-src-snapshot-even-if-not-due flag is specified, or if the most recent scheduled snapshot is somehow missing. In the latter case bzfs immediately creates a snapshot (named with the current time, not backdated to the missed time), and then resumes the original schedule. If the --create-src-snapshot-suffix is '_adhoc' or not a known period then a snapshot is considered non-periodic and is thus created immediately regardless of the creation time of any existing snapshot.

The implementation attempts to fit as many datasets as possible into a single (atomic) 'zfs snapshot' command line, using case-insensitive sort order, and using 'zfs snapshot -r' to the extent that this is compatible with the schedule and --{include|exclude}-dataset* pruning policy. The snapshots of all datasets that fit within the same single 'zfs snapshot' CLI invocation will be taken within the same ZFS transaction group, and correspondingly have identical 'createtxg' ZFS property (but not necessarily identical 'creation' ZFS time property as ZFS actually provides no such guarantee), and thus be consistent. Dataset names that can't fit into a single command line are spread over multiple command line invocations, respecting the limits that the operating system places on the maximum length of a single command line, per `getconf ARG_MAX`.

The baseline implementation uses the 'zfs list -t snapshot' CLI (in parallel) to find the most recent snapshots, which is simple but doesn't scale well with the number of snapshots, at least if the goal is to take snapshots every second. An alternative, much more scalable, implementation queries the standard ZFS "snapshots_changed" dataset property (requires zfs >= 2.2.0), in combination with a local cache that stores this property, as well as the creation time of the most recent snapshot, for each snapshot name pattern and each dataset.

Signed-off-by: Wolfgang Hoschek <wolfgang.hoschek@mac.com>